backend/django_backend/requirements.txt
Django==5.1.2
psycopg2-binary==2.9.10
dj-rest-auth==6.0.0
django-allauth==65.1.0
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
channels==4.1.0
daphne==4.1.2
